More changes to set up tests
