+++ e2e/test/visual/static-visualizations/funnel.cy.spec.js
@@ -1,12 +1,13 @@
 import {
-  restore,
-  setupSMTP,
   openEmailPage,
+  restore,
   sendSubscriptionsEmail,
+  setupSMTP,
   visitDashboard,
 } from "e2e/support/helpers";
 
-import { USERS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
+import { USERS } from "e2e/support/cypress_data";
+import { createFunnelBarQuestion } from "e2e/support/helpers/e2e-visualization-helpers";
 
 const { admin } = USERS;
 
@@ -33,18 +34,3 @@ describe("static visualizations", { tags: "@external" }, () => {
     });
   });
 });
-
-function createFunnelBarQuestion() {
-  const query = {
-    name: `funnel`,
-    native: {
-      query: `SELECT * FROM ( VALUES ('Stage 1', 1000), ('Stage 2', 400), ('Stage 3', 250), ('Stage 4', 100), ('Stage 5', 20), ('Stage 6', 10))`,
-      "template-tags": {},
-    },
-    visualization_settings: {},
-    display: "funnel",
-    database: SAMPLE_DB_ID,
-  };
-
-  return query;
-}
