+++ e2e/test/visual/static-visualizations/gauge.cy.spec.js
@@ -1,15 +1,13 @@
 import {
-  restore,
-  setupSMTP,
   openEmailPage,
+  restore,
   sendSubscriptionsEmail,
+  setupSMTP,
   visitDashboard,
 } from "e2e/support/helpers";
 
-import { USERS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-
-const { ORDERS_ID } = SAMPLE_DATABASE;
+import { USERS } from "e2e/support/cypress_data";
+import { createGaugeQuestion } from "e2e/support/helpers/e2e-visualization-helpers";
 
 const { admin } = USERS;
 
@@ -56,37 +54,3 @@ describe("static visualizations", { tags: "@external" }, () => {
     });
   });
 });
-
-/**
- * @param {number[]} range
- * @param {string[]=} range
- * @param {object=} columnSettings
- */
-function createGaugeQuestion(range, labels, columnSettings) {
-  const colors = ["#ED6E6E", "#F9CF48", "#84BB4C", "#509EE3"];
-  return {
-    name: `Gauge chart with range "${range}"`,
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-    },
-    visualization_settings: {
-      "gauge.segments": range
-        .map((value, index) => {
-          const nextValue = range[index + 1];
-          if (nextValue) {
-            return {
-              min: value,
-              max: nextValue,
-              color: colors[index],
-              label: labels?.[index] || `Label ${index + 1}`,
-            };
-          }
-        })
-        .filter(value => value),
-      ...(columnSettings && { column_settings: columnSettings }),
-    },
-    display: "gauge",
-    database: SAMPLE_DB_ID,
-  };
-}
