+++ e2e/test/visual/static-visualizations/line-area-bar-combo.cy.spec.js
@@ -1,15 +1,17 @@
 import {
-  restore,
-  setupSMTP,
   openEmailPage,
+  restore,
   sendSubscriptionsEmail,
+  setupSMTP,
   visitDashboard,
 } from "e2e/support/helpers";
 
-import { USERS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-
-const { ORDERS_ID, ORDERS, PRODUCTS } = SAMPLE_DATABASE;
+import { USERS } from "e2e/support/cypress_data";
+import {
+  createOneDimensionOneMetricQuestion,
+  createOneDimensionTwoMetricsQuestion,
+  createOneMetricTwoDimensionsQuestion,
+} from "e2e/support/helpers/e2e-visualization-helpers";
 
 const { admin } = USERS;
 
@@ -44,60 +46,3 @@ describe("static visualizations", { tags: "@external" }, () => {
     });
   });
 });
-
-function createOneDimensionOneMetricQuestion(display) {
-  return {
-    name: `${display} one dimension one metric`,
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-      breakout: [["field", ORDERS.CREATED_AT, { "temporal-unit": "month" }]],
-    },
-    visualization_settings: {
-      "graph.dimensions": ["CREATED_AT"],
-      "graph.metrics": ["count"],
-      "graph.show_values": true,
-    },
-    display: display,
-    database: SAMPLE_DB_ID,
-  };
-}
-
-function createOneDimensionTwoMetricsQuestion(display) {
-  return {
-    name: `${display} one dimension two metrics`,
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"], ["avg", ["field", ORDERS.TOTAL, null]]],
-      breakout: [["field", ORDERS.CREATED_AT, { "temporal-unit": "month" }]],
-    },
-    visualization_settings: {
-      "graph.dimensions": ["CREATED_AT"],
-      "graph.metrics": ["count", "avg"],
-      "graph.show_values": true,
-    },
-    display: display,
-    database: SAMPLE_DB_ID,
-  };
-}
-
-function createOneMetricTwoDimensionsQuestion(visualizationType) {
-  return {
-    name: `${visualizationType} one metric two dimensions`,
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-      breakout: [
-        ["field", ORDERS.CREATED_AT, { "temporal-unit": "month" }],
-        ["field", PRODUCTS.CATEGORY, { "source-field": ORDERS.PRODUCT_ID }],
-      ],
-    },
-    visualization_settings: {
-      "graph.dimensions": ["CREATED_AT", "CATEGORY"],
-      "graph.metrics": ["count"],
-      "graph.show_values": true,
-    },
-    display: visualizationType,
-    database: SAMPLE_DB_ID,
-  };
-}
