+++ e2e/test/visual/static-visualizations/progress-bar.cy.spec.js
@@ -1,12 +1,13 @@
 import {
-  restore,
-  setupSMTP,
   openEmailPage,
+  restore,
   sendSubscriptionsEmail,
+  setupSMTP,
   visitDashboard,
 } from "e2e/support/helpers";
 
-import { USERS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
+import { USERS } from "e2e/support/cypress_data";
+import { createProgressBarQuestion } from "e2e/support/helpers/e2e-visualization-helpers";
 
 const { admin } = USERS;
 
@@ -38,20 +39,3 @@ describe("static visualizations", { tags: "@external" }, () => {
     });
   });
 });
-
-function createProgressBarQuestion({ value, goal }) {
-  const query = {
-    name: `progress bar value=${value} goal=${goal}`,
-    native: {
-      query: `SELECT ${value}`,
-      "template-tags": {},
-    },
-    visualization_settings: {
-      "progress.goal": goal,
-    },
-    display: "progress",
-    database: SAMPLE_DB_ID,
-  };
-
-  return query;
-}
