+++ e2e/test/visual/static-visualizations/row.cy.spec.js
@@ -1,15 +1,13 @@
 import {
-  restore,
-  setupSMTP,
   openEmailPage,
+  restore,
   sendSubscriptionsEmail,
+  setupSMTP,
   visitDashboard,
 } from "e2e/support/helpers";
 
-import { USERS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-
-const { ORDERS_ID, ORDERS, PRODUCTS } = SAMPLE_DATABASE;
+import { USERS } from "e2e/support/cypress_data";
+import { createSingleSeriesRowChart } from "e2e/support/helpers/e2e-visualization-helpers";
 
 const { admin } = USERS;
 
@@ -36,22 +34,3 @@ describe("static visualizations", { tags: "@external" }, () => {
     });
   });
 });
-
-function createSingleSeriesRowChart() {
-  return {
-    name: `Single series row chart`,
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-      breakout: [
-        ["field", PRODUCTS.CATEGORY, { "source-field": ORDERS.PRODUCT_ID }],
-      ],
-    },
-    visualization_settings: {
-      "graph.dimensions": ["CATEGORY"],
-      "graph.metrics": ["count"],
-    },
-    display: "row",
-    database: SAMPLE_DB_ID,
-  };
-}
